# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YelpCamp API",
        "version": "1.0.0",
        "description": "A API documentation for campgrounds and reviews"
    },
    "servers": [
        {
            "url": "http://localhost:4000"
        }
    ],
    "paths": {
        "/campgrounds": {
            "get": {
                "summary": "Get a list of all campgrounds",
                "tags": [
                    "Campgrounds"
                ],
                "responses": {
                    "200": {
                        "description": "A list of campgrounds",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Campground"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new campground",
                "tags": [
                    "Campgrounds"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Campground"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Campground created successfully"
                    }
                }
            }
        },
        "/campgrounds/new": {
            "get": {
                "summary": "Render the form to create a new campground",
                "tags": [
                    "Campgrounds"
                ],
                "responses": {
                    "200": {
                        "description": "Render the new campground form"
                    }
                }
            }
        },
        "/campgrounds/search": {
            "get": {
                "summary": "Search for campgrounds by title or location",
                "tags": [
                    "Campgrounds"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "q",
                        "schema": {
                            "type": "string"
                        },
                        "required": false,
                        "description": "Search term for campground title or location"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list of campgrounds matching the search term (Title or Location)"
                    }
                }
            }
        },
        "/campgrounds/{id}": {
            "get": {
                "summary": "Get a specific campground by ID",
                "tags": [
                    "Campgrounds"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The ID of the campground to retrieve"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A specific campground",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Campground"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Update a specific campground by ID",
                "tags": [
                    "Campgrounds"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The ID of the campground to update"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Campground"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Campground updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "campground": {
                                            "$ref": "#/components/schemas/Campground"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific campground by ID",
                "tags": [
                    "Campgrounds"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The ID of the campground to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Campground deleted successfully"
                    }
                }
            }
        },
        "/campgrounds/{id}/edit": {
            "get": {
                "summary": "Render the form to edit a specific campground by ID",
                "tags": [
                    "Campgrounds"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The ID of the campground to edit"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Render the edit campground form",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Campground"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/campgrounds/{id}/reviews": {
            "post": {
                "summary": "Create a new review for a campground",
                "tags": [
                    "Reviews"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The ID of the campground to create a review for"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Review"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Review"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/campgrounds/{id}/reviews/{reviewId}": {
            "delete": {
                "summary": "Delete a specific review from a campground",
                "tags": [
                    "Reviews"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The ID of the campground"
                    },
                    {
                        "in": "path",
                        "name": "reviewId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The ID of the review to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized or forbidden action"
                    },
                    "404": {
                        "description": "Review or campground not found"
                    }
                }
            }
        },
        "/register": {
            "get": {
                "summary": "Render registration page",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Registration page rendered successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "summary": "Register a new user",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/login": {
            "get": {
                "summary": "Render login page",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Login page rendered successfully"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "summary": "Log in a user",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "401": {
                        "description": "Unauthorized – Invalid credentials"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "summary": "Log out the currently logged-in user",
                "tags": [
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "401": {
                        "description": "Unauthorized – User is not logged in"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Image": {
                "type": "object",
                "properties": {
                    "url": {
                        "type": "string"
                    },
                    "filename": {
                        "type": "string"
                    }
                }
            },
            "Campground": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Image"
                        }
                    }
                }
            },
            "Review": {
                "type": "object",
                "properties": {
                    "body": {
                        "type": "string"
                    },
                    "rating": {
                        "type": "number"
                    },
                    "author": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "username": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Campgrounds",
            "description": "API endpoints for managing campgrounds"
        },
        {
            "name": "Reviews",
            "description": "API endpoints for managing reviews"
        },
        {
            "name": "Users",
            "description": "API endpoints for managing users"
        }
    ]
}